# opentsdb: minor optimization to Replace. No alloc on happy path.