Benchmarks:

Before:
```
BenchmarkReplace_Noop-8        	 1000000	      1865 ns/op	     592 B/op	      51 allocs/op
BenchmarkReplace_Something-8   	 1000000	      1442 ns/op	     336 B/op	      37 allocs/op
```

After:
```
BenchmarkReplace_Noop-8        	 5000000	       327 ns/op	       0 B/op	       0 allocs/op
BenchmarkReplace_Something-8   	 1000000	      1676 ns/op	     336 B/op	      37 allocs/op
```

Could potentially optimize the replacing case, especially since we are almost always reducing the size of the string, but I think this should relieve some of the allocs I saw.