@gbrayut 2016-07-27
Testing looks good so far. Gonna merge to master and let the dev systems run overnight